amelia2mitml.list {mitml}	R Documentation

Convert objects of class amelia to mitml.list

Description

This function converts a amelia class object (as produced by the Amelia package) to mitml.list. The resulting object may be used in further analyses.

Usage

amelia2mitml.list(x)

Arguments

x
An object of class amelia as produced by amelia (see the Amelia package).

Value

A list of imputed data sets with an additional class attribute mitml.list.

Author(s)

Simon Grund

See Also

mitmlComplete

Examples

data(studentratings)

require(Amelia)
imp <- amelia(x = studentratings[, c("ID", "MathAchiev", "ReadAchiev")], cs = "ID")

implist <- amelia2mitml.list(imp)

[Package mitml version 0.4-1 Index]