as.mitml.list {mitml}	R Documentation

Convert a list of data sets to mitml.list

Description

This function adds a mitml.list class attribute to a list of data frames. The resulting object can be used in further analyses.

Usage

as.mitml.list(x)

Arguments

x
A list of data frames.

Value

The original list with an additional class attribute mitml.list. The list entries are converted into a data.frame if necessary, in which case a note is printed.

Author(s)

Simon Grund

See Also

is.mitml.list, long2mitml.list

Examples

# data frame with 'imputation' indicator
dat <- data.frame(imputation = rep(1:10, each = 20), x = rnorm(200))

# split into a list and convert to 'mitml.list'
l <- split(dat, dat$imputation)
l <- as.mitml.list(l)

is.mitml.list(l)
# TRUE

[Package mitml version 0.4-1 Index]