confint.mitml.testEstimates {mitml}	R Documentation

Compute confidence intervals

Description

Computes confidence intervals on the basis of the final parameter estimates and inferences given by testEstimates.

Usage

## S3 method for class 'mitml.testEstimates'
confint(object, parm, level = 0.95, ...)

Arguments

object
An object of class mitml.testEstimates as produced by testEstimates.

parm
(optional) A reference to the parameters for which to calculate confidence intervals. Can be a character or integer vector denoting names or position of parameters, respectively. If missing, all parameters are considered (the default).

level
The confidence level. Default is to 0.95 (i.e., 95%).

...
Not being used.

Details

This function computes confidence intervals with the given confidence level for the pooled parameters on the basis of a t-distribution, with estimates, standard errors, and degrees of freedom as returned by testEstimates.

Value

A matrix containing the lower and upper bounds of the confidence intervals.

Author(s)

Simon Grund

See Also

testEstimates

Examples

data(studentratings)

fml <- ReadDis ~ ReadAchiev + (1|ID)
imp <- panImpute(studentratings, formula = fml, n.burn = 500, n.iter = 100, m = 5)

implist <- mitmlComplete(imp)

# fit regression model
fit <- with(implist, lm(ReadDis ~ 1 + ReadAchiev))
est <- testEstimates(fit)

# compute confidence intervals
confint(est)

# ... with different confidence levels
confint(est, level = 0.90)
confint(est, level = 0.999)

[Package mitml version 0.4-1 Index]