is.mitml.list {mitml}	R Documentation

Check if an object is of class mitml.list

Description

This function checks if its argument is a list of class mitml.list.

Usage

is.mitml.list(x)

Arguments

x
An R object.

Value

TRUE or FALSE. A warning message is displayed if the contents of x do not appear to be data frames.

Author(s)

Simon Grund

See Also

as.mitml.list

Examples

l <- list(data.frame(x = rnorm(20)))
l <- as.mitml.list(l)
is.mitml.list(l)
# TRUE

l <- as.list(1:10)
is.mitml.list(l)
# FALSE

class(l) <- "mitml.list"
is.mitml.list(l)
# TRUE, with a warning

[Package mitml version 0.4-1 Index]