write.mitmlSPSS {mitml}	R Documentation

Write mitml objects to SPSS compatible format

Description

Saves objects of class mitml as a text and a syntax file which can be processed by the statistical software SPSS (IBM Corp., 2013).

Usage

write.mitmlSPSS(x, filename, sep = "\t", dec = ".", na.value = -999, syntax = TRUE,
  locale = NULL)

Arguments

x
An object of class mitml or mitml.list as produced by panImpute, jomoImpute, mitmlComplete, or similar).

filename
File base name of the data and syntax files, specified without file extension.

sep
The field separator.

dec
The decimal separator.

na.value
A numeric value coding the missing data in the resulting data file.

syntax
A logical flag indicating if an SPSS syntax file should be generated. This file contains instructions for SPSS for reading in the data file. Default is TRUE.

locale
(optional) A character string specifying the localization to be used in SPSS (e.g., "en_US", "de_DE"; see 'Details').

Details

In SPSS, multiply imputed data are contained in a single file, in which an Imputation_ variable separates the original data and the various imputed data sets. During export, factors are converted to numeric, whereas character variables are left “as is”.

By default, write.mitmlSPSS generates a raw text file containing the data, along with a syntax file containing instructions for SPSS. This syntax file mimics SPSS's functionality to read text files with sensible settings. In order to read in the data, the syntax file must be opened and executed using SPSS, or open using the GUI. Manual changes to the syntax file can be required, for example, if the file path of the data file is not correctly represented in the syntax. The locale argument can be used to ensure that SPSS reads the data in the correct locale.

Alternatively, write.mitmlSAV may be used for exporting directly to the SPSS native .sav format.

Value

None (invisible NULL).

Author(s)

Simon Grund

References

IBM Corp. IBM SPSS Statistics for Windows. Armonk, NY: IBM Corp

See Also

panImpute, jomoImpute, mitmlComplete, write.mitmlSAV

Examples

## Not run: 
data(studentratings)

fml <- ReadDis + SES ~ ReadAchiev + (1|ID)
imp <- panImpute(studentratings, formula = fml, n.burn = 1000, n.iter = 100, m = 5)

# write data file and SPSS syntax
write.mitmlSPSS(imp, filename = "imputation", sep = "\t", dec = ".",
                na.value = -999, locale = "en_US")

## End(Not run)

[Package mitml version 0.4-1 Index]