mids2mitml.list {mitml}	R Documentation

Convert objects of class mids to mitml.list

Description

This function converts a mids class object (as produced by the mice package) to mitml.list. The resulting object may be used in further analyses.

Usage

mids2mitml.list(x)

Arguments

x
An object of class mids as produced by mice (see the mice package).

Value

A list of imputed data sets with class mitml.list.

Author(s)

Simon Grund

See Also

mitmlComplete

Examples

data(studentratings)

# imputation using mice
require(mice)
imp <- mice(studentratings)

implist <- mids2mitml.list(imp)

[Package mitml version 0.4-1 Index]