read.mitml {mitml}	R Documentation

Read mitml objects from file

Description

This function loads mitml class objects from R binary formats (similar to ?load), usually produced by write.mitml.

Usage

read.mitml(filename)

Arguments

filename
Name of the file to read, to be specified with file extension (e.g., .R, .Rdata).

Value

An object of class mitml.

Author(s)

Simon Grund

See Also

panImpute, jomoImpute, write.mitml

Examples

## Not run: 
data(studentratings)

fml <- ReadDis + SES ~ ReadAchiev + (1|ID)
imp <- panImpute(studentratings, formula = fml, n.burn = 1000, n.iter = 100, m = 5)

# write 'mitml' object
write.mitml(imp, filename = "imputation.R")

# read previously saved 'mitml' object
old.imp <- read.mitml("imputation.R")

class(old.imp)
old.imp

## End(Not run)

[Package mitml version 0.4-1 Index]