studentratings {mitml}	R Documentation

Example data set on student ratings and achievement

Description

Contains simulated data for students nested within schools, featuring students' ratings of their teachers' behavior (i.e., disciplinary problems in mathematics and reading class) and their general learning environment (school climate) as well as mathematics and reading achievement scores, and scores for socio-economic status and cognitive ability.

In addition, the data set features the ID of 50 different schools (i.e., clusters), the biological sex of all students, and a broad, additional grouping factor. Different amounts of missing data have been inserted into the data set in a completely random fashion.

Usage

data(studentratings)

Format

A data frame containing 750 observations on 10 variables.

[Package mitml version 0.4-1 Index]