summary.mitml {mitml}	R Documentation

Summary measures for imputation models

Description

Provides summary statistics and additional information on imputations in objects of class mitml.

Usage

## S3 method for class 'mitml'
summary(object, n.Rhat = 3, goodness.of.appr = FALSE, autocorrelation = FALSE, ...)

Arguments

object
An object of class mitml as produced by panImpute or jomoImpute.

n.Rhat
(optional) An integer denoting the number of segments used for calculating the potential scale reduction factor. Default is 3.

goodness.of.appr
(optional) A logical flag indicating if the goodness of approximation should be printed. Default is FALSE (see 'Details').

autocorrelation
(optional) A logical flag indicating if the autocorrelation should be printed. Default is FALSE (see 'Details').

...
Not used.

Details

The summary method calculates summary statistics for objects of class mitml as produced by panImpute or jomoImpute. The output includes the potential scale reduction factor (PSRF, or \hat{R}) and (optionally) the goodness of approximation and autocorrelation.

The PSRF is calculated for each parameter of the imputation model and can be used as a convergence diagnostic (Gelman and Rubin, 1992). Calculation of the PSRFs can be suppressed by setting n.Rhat = NULL. The PSRFs are not computed from different chains but by dividing each chain from the imputation phase into a number of segments as denoted by n.Rhat. This is slightly different from the original method proposed by Gelman and Rubin.

The goodness of approximation measure indicates what proportion of the posterior standard deviation is due to simulation error. This is useful for assessing the accuracy of the posterior summaries (e.g., the EAP). The autocorrelation includes estimates of the autocorrelation in the chains at lag 1 (i.e., for consecutive draws) and for lags k and 2k, where k is the number of iterations between imputations. For lag k and 2k, the autocorrelation is slightly smoothed to reduce the influence of noise on the estimates (see plot.mitml).

Value

An object of class summary.mitml. A print method is used for more readable output.

Author(s)

Simon Grund

References

Gelman, A., and Rubin, D. B. (1992). Inference from iterative simulation using multiple sequences. Statistical Science, 7, 457-472.

Hoff, P. D. (2009). A first course in Bayesian statistical methods. New York, NY: Springer.

See Also

panImpute, jomoImpute, plot.mitml

Examples

data(studentratings)

fml <- ReadDis + SES ~ ReadAchiev + (1|ID)
imp <- panImpute(studentratings, formula = fml, n.burn = 1000, n.iter = 100, m = 5)

# print summary
summary(imp)

[Package mitml version 0.4-1 Index]